air pollution api's are developed

+++ src/main/java/Weather/Application/sai/DTO/CoordinatesDTO.java
@@ -1,11 +1,13 @@
 package Weather.Application.sai.DTO;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class CoordinatesDTO {
-    private long lat;
-    private long lon;
+
+    private double lat;
+    private double lon;
 }
